+++ Engine/include/sge.h
@@ -65,4 +65,5 @@
 #include "component/ShaderComponent.h"
 #include "component/ObjectComponent.h"
 #include "component/NativeScriptComponent.h"
+#include "component/RenderableComponent.h"
 #include "runtime/Prefab.h"
