+++ Engine/src/component/Component.cpp
@@ -33,18 +33,6 @@ void TestComp::attachToEntity(std::shared_ptr<Component> c, Entity entityHandler
 	attachSimple<TestComp>(c, entityHandler);
 }
 
-void RenderableComponent::attachToEntity(std::shared_ptr<Component> c, Entity entityHandler, Scene& scene)
-{
-    (void)scene;
-    attachSimple<RenderableComponent>(c, entityHandler);
-}
-
-//void MeshComponent::attachToEntity(std::shared_ptr<Component> c, Entity entityHandler, Scene& scene)
-//{
-//    (void)scene;
-//    attachSimple<MeshComponent>(c, entityHandler);
-//}
-
 void InstanceBatch::attachToEntity(std::shared_ptr<Component> c, Entity entityHandler, Scene& scene)
 {
         (void)scene;
