+++ Engine/src/component/Component.h
@@ -100,27 +100,6 @@ struct EngineAPI SkyboxComponent : public Component
         static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
 };
 
-
-struct EngineAPI RenderableComponent : public Component
-{
-	RenderableComponent() = default;
-
-	static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
-
-	template <class Archive>
-	void serialize(Archive& archive) {
-		SERIALIZED_MEMBER(renderTechnique);
-	}
-
-	enum class RenderTechnique : int
-	{
-		Forward,
-		Deferred
-	};
-
-    RenderTechnique renderTechnique = RenderTechnique::Deferred;
-};
-
 struct EngineAPI InstanceBatch : public Component
 {
 public:
